Simplify android_sdk dependency.
Previously the different android_sdk modules were downloaded
independently. This was for optimization purposes but we were always
downloading all the modules removing the benefits of the separation.
This change will use the new bundle from a single cipd package with the
benefit of preventing cache pollution when packages are left on the
cache but not used by the next tasks.

Change-Id: I188402f15ebc1924e6512861eed01a00ea7d24d3
Bug: https://github.com/flutter/flutter/issues/83055
Bug: https://github.com/flutter/flutter/issues/82565
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/13780
Reviewed-by: Jonah Williams <jonahwilliams@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>