Build android tests inside a goma context.
Recipes will be rolling a new version of the depot_tools where autoninja
fails if the targets are running outside of a goma context. Furthermore
in order to split builds from tests we need separate the functionality
to build the tests from the actual run of the tests.

Change-Id: I145f8266ee21f4584c87c688b6d90e090ceca4cc
Bug: https://github.com/flutter/flutter/issues/84618
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/14700
Reviewed-by: Dan Field <dnfield@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>