Add a flaky step for engine builders
This makes it easy to identify flakiness from luci build detailed page, and to collect flaky statistics and present in a dashboard. Additionally, we can also use this to trigger a notification when a flake happens.

This cl also moves the flaky_step function to the test module, so that all recipes can share.

Change-Id: I1ca6f43f39c7115f139e108981159ec0c0aa8954
Bug: https://github.com/flutter/flutter/issues/89308
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/18341
Reviewed-by: Zach Anderson <zra@google.com>
Reviewed-by: Godofredo Contreras <godofredoc@google.com>
Commit-Queue: Keyong Han <keyonghan@google.com>