Remove android_sdk recipe module.
This recipe was used to install the android sdk as different components.
This approach caused several issues with bad state caches and the
inability to update specific components without cleaning up caches
first. This functionality was replaced with a single cipd bundle that
includes all the expected components and it is installed as a generic
build dependency.

Change-Id: I33f34994d1f95f6becff08f8e2220001fce27821
Bug: https://github.com/flutter/flutter/issues/83055
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/19061
Reviewed-by: Casey Hillers <chillers@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>