Copy sdk, ssh, and vdl recipes from fuchsia recipes
These recipe modules are only used by flutter, so it makes sense for
them to live in the flutter recipes repo instead of in the fuchsia
recipes repo.

This is also 3 fewer modules that will need to maintain py2
compatibility in the fuchsia recipes repo as long as flutter's migration
to py3 is still in progress.

Bug: https://fxbug.dev/84595
Change-Id: I654a1a12f807914e028539236958515903e93853
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/19100
Reviewed-by: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>