Reland "Use property branch instead of the one from buildbucket".
Context: The change was reverted because manual runs set git_ref with
the commit to be retried. This made the logic to think the commit was
coming from presubmit and the --branch flag was not setup.

This is to allow re-running tasks from production branches.

Change-Id: Ia7e23c37b90f97038b6391fcd58a84abd4580ae1
Bug: https://github.com/flutter/flutter/issues/92127
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/19561
Reviewed-by: Casey Hillers <chillers@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>