Guess the branch for the current stable | beta commit.
On release branches we don't know the branch name the commit is coming
from. In some cases knowing the source branch is required to reliably
use the correct versions of dependent repositories to use. An example of
this is engine builds checking out framework, framework checking out
plugins or adhoc tests checking out the correct versions of the code.

Bug: https://github.com/flutter/flutter/issues/94948
Bug: https://github.com/flutter/flutter/issues/94949
Bug: https://github.com/flutter/flutter/issues/91921
Change-Id: Ide053d33f330871515c04c9aa544a25c178e87bd
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/25660
Reviewed-by: Casey Hillers <chillers@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>