Use autoninja for ios_debug_sim_arm64 builds
This will save ~5-10 minutes on a few builders.

Also, un-parallelize the ios_release builds. It doesn't buy anything.

We'll be able to speed up all the iOS builders a lot when we
drop the 32-bit builds from the main branch in a couple weeks.

Change-Id: Ibd162e91ef853da4d4093e92acb649f6248a34cf
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/28100
Reviewed-by: Godofredo Contreras <godofredoc@google.com>
Commit-Queue: Zach Anderson <zra@google.com>