Refactor packaging/packaging.py
This change removes two out of the three git checkout's from this recipe.
It also fixes a bug where we were running the version of the
prepare_package.dart script from the release branch, rather than from
HEAD.

Compare the number of steps in a led run from this change:
https://ci.chromium.org/raw/build/logs.chromium.org/flutter/led/fujino_google.com/27f117d0ef62005d08fc96c02cfb220b55e5e49920b3327a7a4ef83a11757910/+/build.proto?server=chromium-swarm.appspot.com

With the last successful run in prod:
https://ci.chromium.org/ui/p/flutter/builders/prod/Mac%20Arm64%20Flutter%20Beta%20Packaging/15/overview

Also note the number of lines deleted from the expectations.

Bug: https://github.com/flutter/flutter/issues/102255
Change-Id: Ia7b85559ad5a7f01d4a5363268a686d3e94c77b5
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/29200
Reviewed-by: Casey Hillers <chillers@google.com>
Commit-Queue: Christopher Fujino <fujino@google.com>