Eliminate Windows UWP build
We have decided to end development of the Windows UWP embedder. This
removes the UWP build and upload steps. It also marks the
`build_windows_uwp` field as optional and deprecated in engine.proto in
order for us to transition off using this field. Once all usages have
been removed, we can mark it `reserved` in order to prevent future use.

Tool support was removed in:
https://github.com/flutter/flutter/pull/102174

BUG=https://github.com/flutter/flutter/issues/102172

Change-Id: Idde0508e0e8f06bd7251dc41bd9c63e4489e5c2e
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/29500
Commit-Queue: Chris Bracken <cbracken@google.com>
Reviewed-by: Jonah Williams <jonahwilliams@google.com>
Reviewed-by: Casey Hillers <chillers@google.com>