Do not use publish flag in any case in call to packaging script
Not using the publish flag on the beta|stable branches only uses the change made to the packaging script which still updates metadata but does not publish the archive. Then, the recipe handles uploading and is able to separate the compilation from the uploading stages for BCID compliance.

Change-Id: I6374ace5299f6578d86bbb7509b0609dfb124b6b
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/36680
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Godofredo Contreras <godofredoc@google.com>