Add additional logic to ensure branch is beta|stable
This change adds additional safety in making sure the packaging for the flutter repo is only being triggered on the correct release branches, which are now beta|stable. Formerly, the main branch was allowed for testing, but this change will no longer allow that.

Change-Id: I564dcb83df86bb1efae072598898eac58c886dc7
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/37100
Reviewed-by: Godofredo Contreras <godofredoc@google.com>
Commit-Queue: Jesse Seales <jseales@google.com>