Update to override sha when checking out multiple repoistories with cocoon.
When we currently perform a checkout of cocoon and flutter we get the cocoon sha from the build. We checkout that version of cocoon and then when we go to checkout flutter we end up using that sha so then the flutter checkout fails since that commit in cocoon is not in the flutter repo. This just allows us to override that and use the original ref which for the coming change is just the tot master branch.

Bug: https://github.com/flutter/flutter/issues/119012
Change-Id: Ie422c9c8e22ecbf96e73563dcfa61a3cbf4fb6f3
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/38133
Reviewed-by: Yusuf Mohsinally <mohsinally@google.com>
Commit-Queue: Ricardo Amador <ricardoamador@google.com>