Do not fail packaing tests when running on presubmit.
The current behavior is to fail the builds if the git_ref is not coming
from refs/heads/xyz. Presubmit tests use refs/pull/#/heads which is
making the tests fail.

This CL will print a step message to signal is not uploading rather than
failing.

Change-Id: I991681804ebfb902dde16486c696346b4b5432d4
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/41021
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Drew Roen <drewroen@google.com>