Add android user home for android sdk dep.
Android tests were using the default location which had two different
issues: it was writing state outside of the test chroot, and it was
deleting the state after the test execution. This change sets the
android home to a tmp directory that is deleted after the task is
complete.

Bug: https://github.com/flutter/flutter/issues/126557
Change-Id: I7237b5eaccdfa252930faa5b293bf1cc17fbf937
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/44380
Reviewed-by: Keyong Han <keyonghan@google.com>
Reviewed-by: Ricardo Amador <ricardoamador@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Gray Mackall <mackall@google.com>