Fix xcode cache cleanup
The old logic cleans up xcode cache when xcode path exists but runtime path doesn't. This misses the case where the runtime path exists but missing some files (not able to finish copying).

This CL uses `xcrun simctl list runtimes` as the source of truth of xcode cache pollution. If it outputs no runtime, then we are sure the xcode is installed in an incomplete state.

Change-Id: Ic98dc1fefeb22f03e5090dc7145ace7e0c14e49b
Bug: https://github.com/flutter/flutter/issues/125273
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/44761
Reviewed-by: Ricardo Amador <ricardoamador@google.com>
Commit-Queue: Keyong Han <keyonghan@google.com>