Move firebaselab configs to ci.yaml.
The current implementation forces to update recipes to
add/remove/updated physical or virtual devices from firebaselab tests.
This forces a monolithic tests running all the configurations and
removes the ability to move some configs to bringup.

This change moves the configurations to properties that can be passed
from .ci.yaml. It will allow to split the monolithic test as multiple
builds that can be marked as bringup independently.

Bug: https://github.com/flutter/flutter/issues/127884
Bug: https://github.com/flutter/flutter/issues/118736
Change-Id: I063ce80c7936bfa00fa14473cd253ac4ae46fd89
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/45182
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Reid Baker <reidbaker@google.com>