Assert valid tag and release_channel combo in release MPA workflow.
In order to add an additional layer of validation, this change adds a regex match to ensure that for each release channel (i.e.g 'stable' or 'beta') there is a tag which is formatted appropriately.
After an issue of a tag malformed for the release channel on which it was pushed, this change will make the release workflow fail if the combination is incorrect, and will clearly show the assertion error of the tag not being valid.

Github issue link: https://github.com/flutter/flutter/issues/128792

Change-Id: I59b86e4b0db45666dabc29e8ea8196d9ba05660f
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/46000
Reviewed-by: Casey Hillers <chillers@google.com>
Commit-Queue: Jesse Seales <jseales@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>