Ensure that the ToT engine commit is contained within flutter/flutter candidate.
When releasing, this addition to the publish bot asserts that the commit found in flutter/flutter engine.internal file matches what is found on the same candidate branch in the engine. This is in order to validate that the commit has been rolled into the release candidate successfully.

If the commit validation fails, the release will fail at this assertion.

Bug: 289062103

Change-Id: I0da2682fdfdd1908247397ea3243227959657c97
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/46840
Commit-Queue: Jesse Seales <jseales@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>