Adding support for android emulators for packages.
The emulators can be enabled in the packages targets by adding the dependency for the emulator to the .ci.yaml targets. For example:

properties:
      dependencies: >-
        [
          {"dependency": "android_sdk", "version": "version:33v6"},
          {"dependency": "android_virtual_device", "version": "33"},
          {"dependency": "open_jdk", "version": "version:11"}
        ]
      target_file: android_platform_tests.yaml
      channel: master
      version_file: flutter_master.version
      cores: "32"
      package_sharding: "--shardIndex 0 --shardCount 6"

You may or may not need to include the android_sdk or open_jdk dependencies depending on if the platform targets include these. The only dependency that is required to activate the emulator is:

{"dependency": "android_virtual_device", "version": "33"}

If that is included it is assumed that an emulated device is wanted for testing.


Saving temp changes

Change-Id: Ia8f26edf24e4a232e34995b2875fea7ecd1f7b31
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/47340
Commit-Queue: Ricardo Amador <ricardoamador@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>