Initial commit to add BCID provenance verification.
Given the restricted access to dart-internal, wrap the initial verification in a try-catch so that it is not blocking. This will allow the builder recipe to continue working while we ensure BCID verification is working correctly.

BCID verification will not be considered complete until this try-except is removed, as correctly implemented BCID verification SHOULD fail the build

Bug: b/272046059
Change-Id: I8dac50937bfe36b5462643cb7f77794885533fb6
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/48040
Commit-Queue: Drew Roen <drewroen@google.com>
Reviewed-by: Jesse Seales <jseales@google.com>
Reviewed-by: Godofredo Contreras <godofredoc@google.com>