Fix version collection in Android Emulator tests.
For some reason, which I suspect is how this object is instantiated, the version is not being picked up unless it is passed into the context start method. This is a quick fix to get the recipes working for the devs, while I clean this up.

I have a question open with Robbie Iannuci about the instantiation of these objects.

Bug: https://github.com/flutter/flutter/issues/134588
Change-Id: I22a5f0fe3262dbf3fecb4dc27285997c217c3fe0
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/51120
Reviewed-by: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Ricardo Amador <ricardoamador@google.com>
Commit-Queue: Ricardo Amador <ricardoamador@google.com>