Refactor the packaging recipe
Update the release json last, so that BCID verification can happen first. Otherwise, it is possible that the sdk gets uploaded, the release json is updated (allowing the sdk to be downloaded by users), but BCID verification fails and we have an unsafe artifact released.

Change-Id: I94d06f56d975b672cff21a20c296869f7a926974
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/51360
Reviewed-by: Yusuf Mohsinally <mohsinally@google.com>
Reviewed-by: Jesse Seales <jseales@google.com>
Commit-Queue: Drew Roen <drewroen@google.com>