Adding an uninstall method to make sure that we are leaving nothing behind on the device related to the emulator under test.
Add a method that shows the attached devices throughout the test to make sure we have a clean environment and to eliminate any guessing as to whether or not there may be multiple devices on the machine.

This is part of an effort to attempt to remove as many issues from the run as possible and makes the test more hermetic by cleaning up everything relate to the emulator.

Test runs:
https://ci.chromium.org/b/8759249601146257521
https://ci.chromium.org/b/8759250069999315489

I have run this upwards of 20 times and there have been no failures. I suspect there may be some when running in the CI which hopefully this will help to highlight any issues there.

Bug: Part of https://github.com/flutter/flutter/issues/141292
Change-Id: I498182fa05c0530a47513cfe05e439e7bfefea70
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/53903
Reviewed-by: Yusuf Mohsinally <mohsinally@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>
Commit-Queue: Ricardo Amador <ricardoamador@google.com>