[flutter_deps] Remove stray backslash from example recipe
Newer versions of the path recipe module don't allow backslashes in path strings. There are other references to a "flutter sdk" directory, so the backslash was probably a typo. Given that this is just a change to an example recipe, it should be harmless.

Original CL that introduced the type mentions space but no backslash:
https://flutter-review.googlesource.com/c/recipes/+/8360

Change-Id: I1fb3a69cfe2a1a5548a62cebc5ce9b5b1af40787
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/59322
Reviewed-by: Christopher Fujino <fujino@google.com>