Remove seemingly unused `engine.proto` file.
I think this file was originally used in the pre-engine_v2 builders, and
then was forked to become recipes/release/release.proto, and now is
unused. The way I determined it was unused was by grepping for
"PB.recipes.engine", which is the pattern roughly that matches for
protos that are being used.

Change-Id: Iaee38cff22b1228e43e110826ec6ab3e90d0c2f4
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/64940
Commit-Queue: Matan Lurey <matanl@google.com>
Reviewed-by: Jason Simmons <jsimmons@google.com>