Try adding WWDRCAG3 to login.keychain.
Codesigning is failing due to "unable to build chain to self-signed
root". This usually indicates that the WWDRCAG3 cert is missing. We
previously tried adding it to the build.keychain, since that's the
keychain we use for codesigning, but the error persists. This adds it to
the login.keychain to see if that fixes the issue.

Bug: https://github.com/flutter/flutter/issues/168342, b/417754906
Change-Id: I1e835086757c81ae3e34080da612202b58a90c93
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/65480
Reviewed-by: Matan Lurey <matanl@google.com>
Commit-Queue: Victoria Ashworth <vashworth@google.com>